+++ master_parse/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log for Master Parse Tool
 
+Version 1.13.0:
+
+* Added support for `AZURE_ONLY` and `AMAZON_ONLY` tags, triggered
+  by `--target-profile` command line argument.
+
 Version 1.12.3:
 
 * Added border around video player.